android: native modules support
Adds native modules support to Android builds by including new
gradle tasks to copy the nodejs-project and build the npm native
modules for each of the needed CPU architectures. The resulting
binaries are then added as a different asset folder for each
architecture, to be copied and loaded at runtime depending on which
architecture is needed.
The npm tasks can be activated by setting the environment variable
'NODEJS_MOBILE_BUILD_NATIVE_MODULES' to 1.